+++ src/branches/2.0/modules/topaddrCC/data.h
@@ -43,8 +43,8 @@ typedef struct topaddr_config topaddr_config_t;
 como_tuple struct topaddr_tuple {
     timestamp_t ts;     /* timestamp */
     uint32_t addr;  	/* src/dst address */ 
-    uint64_t bytes;	/* number of bytes */
-    uint32_t pkts;	/* number of packets */
+    double bytes;	/* number of bytes */
+    double pkts;	/* number of packets */
 };
 
 como_record struct topaddr_record {
